requirements.txt
streamlit
langchain
sentence-transformers
langchain-community
sqlite-vss
rank_bm25
translators
openai
mecab-python3
unidic
pyahocorasick